refactor(frontend): Create folder for Lucide icons
