+++ src/frontend/src/lib/components/icons/lucide/IconChangelog.svelte
