+++ src/frontend/src/lib/components/icons/lucide/IconGlasses.svelte
