+++ src/frontend/src/lib/components/icons/lucide/IconInfinity.svelte
