+++ src/frontend/src/lib/components/icons/lucide/IconLogin.svelte
