+++ src/frontend/src/lib/components/icons/lucide/IconManage.svelte
