+++ src/frontend/src/lib/components/icons/lucide/IconShieldAlert.svelte
