+++ src/frontend/src/lib/components/navigation/ChangelogLink.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-	import IconChangelog from '$lib/components/icons/IconChangelog.svelte';
+	import IconChangelog from '$lib/components/icons/lucide/IconChangelog.svelte';
 	import ExternalLink from '$lib/components/ui/ExternalLink.svelte';
 	import { i18n } from '$lib/stores/i18n.store';
 	import { replaceOisyPlaceholders } from '$lib/utils/i18n.utils';
