+++ src/frontend/src/lib/components/tokens/TokensMenu.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
 	import { Popover } from '@dfinity/gix-components';
 	import { erc20UserTokensNotInitialized } from '$eth/derived/erc20.derived';
-	import IconManage from '$lib/components/icons/IconManage.svelte';
+	import IconManage from '$lib/components/icons/lucide/IconManage.svelte';
 	import ManageTokensMenuButton from '$lib/components/manage/ManageTokensMenuButton.svelte';
 	import TokensZeroBalance from '$lib/components/tokens/TokensZeroBalance.svelte';
 	import TokensZeroBalanceCheckbox from '$lib/components/tokens/TokensZeroBalanceCheckbox.svelte';
